requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra-index-url=https://download.pytorch.org/whl/cpu requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cpu

absl-py==2.1.0
    # via tensorboard
anyio==4.3.0
    # via
    #   httpx
    #   jupyter-server
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.14.0
    # via jupyterlab-server
beautifulsoup4==4.12.3
    # via nbconvert
biopython==1.83
    # via -r requirements.in
bleach==6.1.0
    # via nbconvert
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.1
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==2.0.1
    # via stack-data
fastjsonschema==2.19.1
    # via nbformat
filelock==3.14.0
    # via torch
fonttools==4.50.0
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
fsspec==2024.3.1
    # via torch
grpcio==1.64.0
    # via tensorboard
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via jupyterlab
idna==3.6
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==6.29.3
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
    #   qtconsole
ipython==8.22.2
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==8.1.2
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.1
    # via ipython
jinja2==3.1.3
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   torch
joblib==1.3.2
    # via scikit-learn
json5==0.9.24
    # via jupyterlab-server
jsonpointer==2.4
    # via jsonschema
jsonschema[format-nongpl]==4.21.1
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==8.6.1
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
    #   qtconsole
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
jupyter-events==0.10.0
    # via jupyter-server
jupyter-lsp==2.2.4
    # via jupyterlab
jupyter-server==2.13.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.1.5
    # via notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.25.4
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.10
    # via ipywidgets
kiwisolver==1.4.5
    # via matplotlib
markdown==3.6
    # via tensorboard
markupsafe==2.1.5
    # via
    #   jinja2
    #   nbconvert
    #   werkzeug
matplotlib==3.8.3
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==3.0.2
    # via nbconvert
mpmath==1.3.0
    # via sympy
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.3
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.3
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.3
    # via torch
notebook==7.1.2
    # via jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==1.26.4
    # via
    #   -r requirements.in
    #   biopython
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tensorboard
    #   torchvision
overrides==7.7.0
    # via jupyter-server
packaging==24.0
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
    #   qtconsole
    #   qtpy
pandas==2.2.1
    # via
    #   -r requirements.in
    #   seaborn
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.2.0
    # via
    #   matplotlib
    #   torchvision
platformdirs==4.2.0
    # via jupyter-core
prometheus-client==0.20.0
    # via jupyter-server
prompt-toolkit==3.0.43
    # via
    #   ipython
    #   jupyter-console
protobuf==5.27.0
    # via tensorboard
psutil==5.9.8
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-json-logger==2.0.7
    # via jupyter-events
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via jupyter-events
pyzmq==25.1.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   qtconsole
qtconsole==5.5.1
    # via jupyter
qtpy==2.4.1
    # via qtconsole
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.31.0
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.4.1.post1
    # via -r requirements.in
scipy==1.12.0
    # via scikit-learn
seaborn==0.13.2
    # via -r requirements.in
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
    #   tensorboard
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
sympy==1.12
    # via torch
tensorboard==2.16.2
    # via -r requirements.in
tensorboard-data-server==0.7.2
    # via tensorboard
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
threadpoolctl==3.4.0
    # via scikit-learn
tinycss2==1.2.1
    # via nbconvert
torch==2.3.0+cpu
    # via
    #   -r requirements.in
    #   torchvision
torcheval==0.0.7
    # via -r requirements.in
torchvision==0.18.0+cpu
    # via -r requirements.in
tornado==6.4
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.66.4
    # via -r requirements.in
traitlets==5.14.2
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   qtconsole
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.11.0
    # via
    #   torch
    #   torcheval
tzdata==2024.1
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.7.0
    # via jupyter-server
werkzeug==3.0.3
    # via tensorboard
widgetsnbextension==4.0.10
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
